refactor: rename module models/custom_user.py to model/user.py
