+++ blog_app/models/__init__.py
@@ -7,7 +7,7 @@ __all__ = [
     'Like',
     'Tag']
 
-from blog_app.models.custom_user import CustomUser
-from blog_app.models.custom_user import TimeStampModel
+from blog_app.models.user import CustomUser
+from blog_app.models.user import TimeStampModel
 from blog_app.models.contact_us import ContactUs
 from blog_app.models.post import Post, Comment, Like, Tag
